Refactoring: GridCreatingConditionTriangle::FinishPolygonDefiningCommand implementation is separated from gridcreatingconditiontriangle.cpp.
