+++ libs/gridcreatingcondition/triangle/triangle.pro
@@ -106,7 +106,8 @@ HEADERS += gcc_triangle_global.h \
            gridcreatingconditiontriangleremeshdialog.h \
            gridcreatingconditiontriangleremeshpolygon.h \
            gridcreatingconditiontrianglesettingdialog.h \
-           private/gridcreatingconditiontriangle_addremeshpolygoncommand.h
+           private/gridcreatingconditiontriangle_addremeshpolygoncommand.h \
+           private/gridcreatingconditiontriangle_finishpolygondefiningcommand.h
 FORMS += gridcreatingconditiontrianglepolygoncoordinateseditdialog.ui \
          gridcreatingconditiontrianglepolylinecoordinateseditdialog.ui \
          gridcreatingconditiontriangleremeshdialog.ui \
@@ -123,7 +124,8 @@ SOURCES += gridcreatingconditioncreatortriangle.cpp \
            gridcreatingconditiontriangleremeshdialog.cpp \
            gridcreatingconditiontriangleremeshpolygon.cpp \
            gridcreatingconditiontrianglesettingdialog.cpp \
-           private/gridcreatingconditiontriangle_addremeshpolygoncommand.cpp
+           private/gridcreatingconditiontriangle_addremeshpolygoncommand.cpp \
+           private/gridcreatingconditiontriangle_finishpolygondefiningcommand.cpp
 TRANSLATIONS += languages/iricGccTriangle_es_ES.ts \
                 languages/iricGccTriangle_fr_FR.ts \
                 languages/iricGccTriangle_id_ID.ts \
